stage analyse execution pipeline
